moved IndexerPlugin to index package
