+++ kifi-backend/modules/search/app/com/keepit/search/index/Indexer.scala
@@ -12,7 +12,6 @@ import org.apache.lucene.index.Term
 import com.keepit.common.db.{SequenceNumber, Id}
 import com.keepit.common.logging.Logging
 import com.keepit.common.time._
-import com.keepit.search.IndexManager
 import com.keepit.search.Searcher
 import play.modules.statsd.api.Statsd
 import org.apache.commons.io.FileUtils
