+++ kifi-backend/modules/search/app/com/keepit/search/message/MessageIndexerPlugin.scala
@@ -5,10 +5,10 @@ import com.keepit.common.actor.ActorInstance
 import com.keepit.common.healthcheck.AirbrakeNotifier
 import com.keepit.common.plugin.SchedulingProperties
 import com.keepit.common.zookeeper.ServiceDiscovery
-import com.keepit.search.IndexerActor
-import com.keepit.search.IndexerPlugin
-import com.keepit.search.IndexerPluginImpl
-import com.keepit.search.IndexManager
+import com.keepit.search.index.IndexerActor
+import com.keepit.search.index.IndexerPlugin
+import com.keepit.search.index.IndexerPluginImpl
+import com.keepit.search.index.IndexManager
 
 trait MessageIndexerPlugin extends IndexerPlugin[MessageIndexer]
